Se corrigen algunos métodos para cumplir con el primer epic bonus

+++ src/main/java/ar/com/ada/api/monitor/models/response/MuestraAlturaMinima.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public class MuestraAlturaMinima {
     public String color;
-    public int alturaNivelDelMarMinima;
+    public double alturaNivelDelMarMinima;
     public Date horario;
     //¿O debería guardarlo como un String? Formato:  “2020-08-04T10:02:02”
